+++ extensions/in-memory/negotiation-store-memory/src/test/java/org/eclipse/dataspaceconnector/negotiation/store/memory/TestFunctions.java
@@ -28,20 +28,29 @@ import java.util.UUID;
 
 public class TestFunctions {
 
-    public static ContractNegotiation createNegotiation(String name) {
-        return ContractNegotiation.Builder.newInstance()
-                .type(ContractNegotiation.Type.CONSUMER)
-                .id(name)
+    public static ContractNegotiation createNegotiation(String id) {
+        return createNegotiationBuilder(id)
                 .contractAgreement(createAgreement())
+                .build();
+    }
+
+    public static ContractNegotiation.Builder createNegotiationBuilder(String id) {
+        return ContractNegotiation.Builder.newInstance()
+                .id(id)
+                .type(ContractNegotiation.Type.CONSUMER)
                 .contractOffers(List.of(ContractOffer.Builder.newInstance().id("contractId")
                         .policy(Policy.Builder.newInstance().build()).build()))
                 .counterPartyAddress("consumer")
                 .counterPartyId("consumerId")
-                .protocol("ids-multipart")
-                .build();
+                .protocol("ids-multipart");
     }
 
     public static ContractAgreement createAgreement() {
+        return createAgreementBuilder()
+                .build();
+    }
+
+    public static ContractAgreement.Builder createAgreementBuilder() {
         return ContractAgreement.Builder.newInstance()
                 .id("agreementId")
                 .providerAgentId("provider")
@@ -50,7 +59,6 @@ public class TestFunctions {
                 .policy(Policy.Builder.newInstance().build())
                 .contractStartDate(Instant.now().getEpochSecond())
                 .contractEndDate(Instant.now().plus(1, ChronoUnit.DAYS).getEpochSecond())
-                .contractSigningDate(Instant.now().getEpochSecond())
-                .build();
+                .contractSigningDate(Instant.now().getEpochSecond());
     }
 }