+++ extensions/sql/contract-negotiation-store/src/test/java/org/eclipse/dataspaceconnector/sql/contractnegotiation/TestFunctions.java
@@ -44,7 +44,6 @@ public class TestFunctions {
         return ContractNegotiation.Builder.newInstance()
                 .type(ContractNegotiation.Type.CONSUMER)
                 .id(id)
-                .contractAgreement(null)
                 .contractAgreement(agreement)
                 .correlationId("corr-" + id)
                 .counterPartyAddress("consumer")